Fix bootstrap-servers with invalid api_interface
When bootstrapping a host for kolla-ansible the 'kolla-ansible
bootstrap-servers' command can fail if for any of the hosts the
'api_interface' is invalid. This could be true for a number of reasons:

- The 'api_interface' variable is undefined.
- The host does not have a network interface matching 'api_interface'.

Valid causes for the second point include:
- The Ansible control host (localhost) is in the inventory but is not
  intended to be a target host.
- One of the hosts (currently must be localhost) is a Bifrost host,
  which does not need a valid api_interface.
- The services mapped to the host do not require the use of an
  api_interface.

This change checks for a valid api_interface variable for the host and
its presence in the host's hostvars before accessing the variable.
This results in successful completion of the command, with the omission
of host entries for those hosts without a valid api_interface.

Change-Id: Ie111ef54130adf2556ce83c402cdbb5058ace4f6
Closes-bug: #1665364