Set any_errors_fatal for all plays
Currently we are using the default value of 'false' for
'any_errors_fatal'.  This means that plays will execute until the end or
all hosts have failed. A host can fail and others may continue. This can
lead to odd situations, and makes errors harder to diagnose.

One specific example of where this can be an issue is during upgrade. If
one host fails early on, then it will not be upgraded. For clustered
services such as RabbitMQ or MariaDB this can cause problems. Many
OpenStack services have specific instructions for rolling upgrades which
would not work well if one or more hosts are not involved.

This change sets any_errors_fatal to true for all plays that are not
just executing against localhost (in which case there is only one host).

Change-Id: I513ad1a0ad770a249f573be8dd47e9ada6787a7a
Closes-Bug: #1833737