Wait for mariadb to stop after shutdown
Stein only.

Currently the kolla-ansible-centos-source-upgrade-ceph job is failing on
the stable/stein branch.

The problem occurs with mariadb, when performing an upgrade to the Stein
release which has a new version of mariadb. It appears that when the
slave mariadb services are shut down, we do not wait for the container
to stop, so the service may not shut down cleanly. This prevents it from
starting up successfully.

This change waits for the container to stop after the shutdown command
has been executed. It also temporarily replaces the restart policy of
the container to prevent it from starting up again after the shutdown.

This is not required in other branches since the mariadb shutdown
workaround was only added in the stein branch for bug 1820325.

There is a second issue that is addressed here. The Stein release
switched from using xtrabackup to mariabackup for galera state syncing.
If we run both container versions at the same time on different hosts
then we can get an error such as the following:

sh: wsrep_sst_mariabackup: command not found

We therefore now stop the cluster and perform a recovery during an
upgrade, if we detect that xtrabackup is in use.

Change-Id: I513bcf31adaee8441d43c6b578ca06f66820e52b
Closes-Bug: #1834191
Related-Bug: #1820325