CentOS 8: Support Elasticsearch & Kibana 6.x
Train only.

Elasticsearch and Kibana 5.x do not support CentOS 8. Kolla Train release
provides elasticsearch6 and kibana6 images which contain Elasticsearch and
Kibana 6.x. These images are supported on CentOS 7 and 8 only. This provides a
path for migration from CentOS 7 to 8.

This change adds 'elasticsearch_use_v6' and 'kibana_use_v6' flags which
can be set to 'true' to deploy Elasticsearch and Kibana 6.x on CentOS 7
or 8. These flags are 'true' on CentOS 8, and 'false' elsewhere. If
upgrading from 5.x to 6.x, this should be done via 'kolla-ansible
upgrade'.

Change-Id: I556ad6a206107a97ab5d5942d034f1a460d761d2
Partially-Implements: blueprint centos-rhel-8